output-tg/requirements.txt
python-telegram-bot==20.0
python-dotenv==0.19.2
aio-pika>=9.3,<11.0
